+++ neo/sys/sys_public.h
@@ -484,9 +484,9 @@ bool			Sys_UnlockMemory( void *ptr, int bytes );
 void			Sys_SetPhysicalWorkMemory( int minBytes, int maxBytes );
 
 // DLL loading, the path should be a fully qualified OS path to the DLL file to be loaded
-int				Sys_DLL_Load( const char *dllName );
-void *			Sys_DLL_GetProcAddress( int dllHandle, const char *procName );
-void			Sys_DLL_Unload( int dllHandle );
+intptr_t		Sys_DLL_Load( const char *dllName );
+void *			Sys_DLL_GetProcAddress( intptr_t dllHandle, const char *procName );
+void			Sys_DLL_Unload( intptr_t dllHandle );
 
 // event generation
 void			Sys_GenerateEvents();
